Assistant/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-dotenv==1.0.1
requests==2.32.3
